+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 
-# Revision history for shrinkage
+# Revision history for covariance
 
 
 ## Unreleased changes
 
 
 ## 0.1.0.0
+
+-   Shrinkage based estimator by Ledoit-Wolf.